Support for reserved VM Classes
This patch adds support for reserving N slots for VM Classes in a
given namespace.

Please note that while there are multiple zones in a namespace, the
current design does not allow a VM using a reserved VM Class to be
scheduled in a zone with no reserved slots available.