Reconcile VMs on async signal
This patch adds support for reconciling VMs when their state has changed
on the underlying platform, ex. vSphere. There are three primary
components:

* the watcher -- emits results on a channel monitored by...
* the service -- receives results and enqueues reconcile requests
* the zone controller -- adds/removes zone folders to/from the watcher

Please note, the environment variable `ASYNC_SIGNAL_DISABLED` may be set
to a truth-y string value, ex. "true", to completely disable the async
signal logic, regardless of the feature state switch.

-~= WATCHER =~-

The watcher is located in `pkg/util/vsphere/watcher` and watches one or
more vSphere entities that can contain VMs, ex. a `Folder`,
`ClusterComputeResource`, `HostSystem`, etc.. The watcher is initialized
with a set of these entities and creates a `ContainerView` for each.
These are added to the watcher's `ListView`, which enables entities to
be added/removed later while the watcher is running. The watcher is
signaled when a VM enters the view of the watcher or when a VM has a
change to one of the on the following properties:

- `config.extraConfig`         -- Signal when the guest changes
                                  something in `guestinfo`.
- `guest.ipStack`              -- Signal on changes to the network
                                  state.
- `guest.net`                  -- Signal on changes to the network
                                  state.
- `summary.config.name`        -- A way to detect when the VM enters the
                                  scope of the watch.
- `summary.guest`              -- Signal on any changes to the guest.
- `summary.overallStatus`      -- Signal on changes to the VM's status.
- `summary.runtime.host`       -- Signal when the host on which a VM is
                                  running has been changed.
- `summary.runtime.powerState` -- Signal when the VM's power state is
                                  changed.
- `summary.storage.timestamp`  -- Signal when the VM's storage
                                  information (capacity, used, etc.) has
                                  changed.

Not all changes to extraConfig cause the watcher to emit a result. The
following extraConfig keys are ignored:

- backupapi.AdditionalResourcesYAMLExtraConfigKey
- backupapi.BackupVersionExtraConfigKey
- backupapi.ClassicDiskDataExtraConfigKey
- backupapi.DisableAutoRegistrationExtraConfigKey
- backupapi.EnableAutoRegistrationExtraConfigKey
- backupapi.PVCDiskDataExtraConfigKey
- backupapi.VMResourceYAMLExtraConfigKey
- "govcsim"
- "guestinfo.metadata"
- "guestinfo.metadata.encoding"
- "guestinfo.userdata"
- "guestinfo.userdata.encoding"
- "guestinfo.vendordata"
- "guestinfo.vendordata.encoding"
- "vmservice.namespacedName"

Additional keys may be ignored as well, but these are ignored by
default in order to prevent an infinite loop:

- VM controller receives reconcile request -->
- VM controller reconfigures VM -->
- VM Op reconfigures VM -->
- watcher sees change -->
- watcher emits result -->
- service sees result -->
- service enqueues reconcile request -->
- VM controller receives reconcile request, causing a loop.

When the watcher notices a VM enter its view or with a change, the
watcher must get the namespace and name for the VM. This happens one of
three ways:

1. The change itself includes the namespace and name in the extraConfig
2. The watcher looks up the namespace and name from the manager's cache,
   where the field `status.uniqueID` is now an indexed field.
3. Finally, the watcher retrieves the property
   `config.extraConfig["vmservice.namespacedName"]` from the vSphere
   server.

If the namespace and name can be determined, the watcher checks if the
VM already exists in Kubernetes with a status.uniqueID field and if the
update type was the VM entering the view of the watcher. If these
conditions are met, no result is emitted for this VM. This prevents
double-reconciling VMs when the Controller-Manager starts up for the
first time. During start-up, the Controller-Manager automatically
reconciles all objects watched by controllers. Since all VMs would also
be entering the view of watchers, this would cause a large-scale double-
reconcile. Therefore, this logic skips emitting results on startup for
VMs that are already deployed.

If the namespace and name are non-empty, the update types was `Enter`
and the VM has an empty `status.uniqueID` field or the update type was
`Modify`, the watcher emits a result on a channel watched by the next
component, the service.

-~= SERVICE =~-

The service is located in `services/vm-watcher` and is responsible for:

* keeping an instance of the watcher running
* receiving results from the watcher and enqueuing reconcile requests

The service will always start a new instance of the watcher as long as
the reason the previous instance failed was due to a login/auth error.
This is to handle the case of credential rotation.

The service starts a watcher with an initial set of entities to watch
that includes the ManagedObject ID for each `Folder` that can contain
VM Service VMs. These folder IDs are gathered by listing all `Zone`
resources on the cluster and collecting the value of
`spec.managedVMs.folderMoID`.

The service monitors results from the watcher. Upon receiving a result,
the service determines if the reported VM is valid, and if so, a
reconcile request is enqueued.

-~= ZONE CONTROLLER =~-

The zone controller is located in `controllers/infra/zone` and reconciles
`topologyv1.Zone` resources.

When a zone resource without a deletion timestamp is reconciled, the
controller adds a finalizer to it and adds the zone's vm service folder
to the list of the entities monitored by the watcher.

When a zone resource with a non-zero deletion timestamp is reconciled,
the controller removes the zone's vm service folder from the list of the
entities monitored by the watcher and removes the finalizer.