Remove lock on vm watcher
This patch removes the lock from the vm watcher, making it more
performant. There is no need for a lock anymore since the vim
objects are not set to nil when the watcher is closed. They will
still be garbage collected when the watcher itself goes out of
scope.

This patch also explicitly defines the ASYNC_SIGNAL_DISABLED
env var in the deployment YAML to make it clear how to disable
this feature.