Do not patch a zone with finalizer if it is being deleted
API server rejects Patch requests when the zone controller
is attempting to add a finalizer to the zone, causing the watcher to
never start.  And since the watcher is the entity removing the
finalizer, the zone stays stuck in deletion forever.  This change
skips a zone from being patched if the zone is being deleted.

Additionally, do not watch folders backing a zone if the zone is
marked for deletion and the zone does not contain the finalizer.  This
can happen in situations where the zone controller has already removed
the finalizer after stopping the watcher, but the zone is somehow not removed.